gloabl loading spinner. fixes before gloabl auth and stripe subscription fix
