+++ package.json
@@ -23,7 +23,7 @@
     "framer-motion": "^11.3.21",
     "lint": "^1.1.2",
     "lucide-react": "^0.429.0",
-    "next": "14.2.5",
+    "next": "^14.2.5",
     "react": "^18.3.1",
     "react-dom": "^18",
     "react-firebase-hooks": "^5.1.1",
