+++ src/components/ui/LoadingContext.tsx
@@ -1,0 +1,28 @@
+'use client';
+
+import React, { createContext, useState, useContext } from 'react';
+
+interface LoadingContextType {
+  isLoading: boolean;
+  setIsLoading: (isLoading: boolean) => void;
+}
+
+const LoadingContext = createContext<LoadingContextType | undefined>(undefined);
+
+export const LoadingProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  const [isLoading, setIsLoading] = useState(false);
+
+  return (
+    <LoadingContext.Provider value={{ isLoading, setIsLoading }}>
+      {children}
+    </LoadingContext.Provider>
+  );
+};
+
+export const useLoading = () => {
+  const context = useContext(LoadingContext);
+  if (context === undefined) {
+    throw new Error('useLoading must be used within a LoadingProvider');
+  }
+  return context;
+};
